Naive open addressing hash map implementation, memory usage and layout optimization

+++ src/temporary_storage.h
@@ -1,6 +1,7 @@
 #include <cstddef>
+#include "common.h"
 
 void init_temporary_storage();
 void clear_temporary_storage();
-void* talloc(size_t size);
-void* trealloc(void* pointer, size_t previous_size, size_t new_size);
+void* talloc(u32 size);
+void* trealloc(void* pointer, u32 previous_size, u32 new_size);
